+++ demo/src/main/resources/static/css/inicio_styles.css
@@ -1,0 +1,125 @@
+* {
+    box-sizing: border-box;
+}
+
+body {
+    margin: 0;
+    height: 100vh; /* Ocupa el 100% de la altura de la ventana del navegador */
+    width: 100%;
+    margin: 0;
+}
+
+header {
+    margin: 0;
+    width: 100%;
+    height: 20%;
+    display: flex;
+    flex-direction: row;
+    background-color: #9F86C0;
+
+}
+
+.columns-nav {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+}
+.columns-nav .first-nav {
+    display: flex;
+    flex-direction: row;
+    justify-content: right;
+    align-items: end;
+    width: 100%;
+    height: 70%;
+    padding: 0;
+}
+.columns-nav .datos-vet {
+    width: 25%;
+    height: 100%;
+    margin: 0px;
+    padding: 0px;
+    border-bottom: #231942 2px solid;
+}
+.datos-vet h1 {
+    text-align: center;
+    padding: 25px 10px 0px 10px;
+    color: #231942d8;
+    font-size: 1.4rem;
+    line-height: 1.3rem;
+    font-family: 'Karla', sans-serif;
+    margin: 0;
+}
+.columns-nav .second-nav {
+    width: 100%;
+    height: 30%;
+    padding: 10px 10px 10px 30px;
+}
+.header-logo {
+    width: 12%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+}
+.header-logo img {
+    border: 2px solid #5E548E; /* Borde claro usando el color #9F86C0 */
+    box-shadow: 0 0 10px #9F86C0; /* Sombra en el borde para un efecto suave */
+    border-radius: 50%;
+    object-fit: cover;
+    transform: translateX(-10%);
+}
+
+
+/** Estilos de lista general **/
+ul {
+    display: flex;
+    flex-direction: row;
+    justify-content: right;
+    width: 60%;
+    height: 37px;
+    margin: 0px;
+    padding: 0px;
+    list-style: none;
+    text-decoration: none;
+    border-bottom: #231942 2px solid;
+}
+li {
+    display: flex;
+    align-items: center;
+    text-align: center;
+    font-size: 21px;
+    width: 160px;
+    height: 35px;
+    font-family: 'Lexend', sans-serif;
+    font-weight: bold;
+}
+a {
+    color: #ffffff;
+    width: 160px;
+    height: 35px;
+    text-decoration: none;
+}
+ul li #active {
+    color:#524788;
+    width: 160px;
+    height: 35px;
+    text-decoration: none;
+}
+#active:hover,
+ul li a:hover {
+    color:#231942;
+    font-size: 22px;
+    transition: all 0.3s ease;
+}
+/** Estilos del main donde se encuentra la información **/
+
+main {
+    width: 100%;
+    height: 80%;
+    background-color: #ffffff;
+}
+
+
+
+
