+++ demo/src/main/resources/static/css/login_styles.css
@@ -102,7 +102,7 @@ form {
     border: 1px solid #ced4da;
     font-family: Arial, sans-serif;
     border-radius: 5px;
-    font-size: 0.75rem;
+    font-size: 0.8rem;
     outline: none;
     transition: border-color 0.2s ease, box-shadow 0.2s ease;
 }
