+++ demo/src/main/resources/templates/Home_veterinario.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" th:href="@{../css/inicio_styles.css}">
+
+    <!-- Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=ABeeZee:ital@0;1&family=Baloo+2:wght@400..800&family=Glass+Antiqua&family=Karla:ital,wght@0,200..800;1,200..800&family=Lexend:wght@100..900&family=Turret+Road:wght@200;300;400;500;700;800&display=swap" rel="stylesheet">
+</head>
+<body>
+    <header>
+        <div class="columns-nav">
+            <nav class="first-nav">
+                <ul>
+                    <li><a href="#">Tratamientos</a></li>
+                    <li><a href="#">Mascotas</a></li>
+                    <li><a href="#">Clientes</a></li>
+                    <li><a id="active" href="#">Veterinario</a></li>
+                </ul>
+
+                <div class="datos-vet">
+                    <h1>Dr. Lorena García Garavito Suarez</h1>
+                </div>
+            </nav>
+
+            <nav class="second-nav">
+
+            </nav>
+        </div>
+
+        <div class="header-logo">
+            <img th:src="@{../images/persona.png}" alt="foto" width="100" height="100">
+        </div>
+
+    </header>
+    <main>
+        
+
+    </main>
+</body>
+</html>
